requirements.txt
langchain
langchain-openai
openai
python-dotenv
pypdf
pytesseract
pdf2image
Pillow
faiss-cpu